refactor(tasks): extract date filter helpers
